+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.7.1] - 2026-2-15
+
+### Changed
+
+-   Updated FlowField Navigation to v0.2.1
+
 ## [0.7.0] - 2026-1-3
 
 ### Added